backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "composio>=0.8.8",
    "fastmcp>=2.11.2",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-socketio>=5.5.1",
    "google-api-python-client>=2.178.0",
    "google-auth-oauthlib>=1.2.2",
    "openai>=1.99.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "uv>=0.8.4",
]
